refactor:: move popup in features && create types in sared/lib
