+++ src/features/popup-gallery/popup-gallery.module.scss
