+++ src/shared/lib/types/index.ts
@@ -1,1 +1,2 @@
 export type { LayoutProps } from './provider'
+export type { BookImage, BookData } from './bookCard'
