+++ src/widgets/book-card/ui/book-card/book-card.tsx
@@ -1,9 +1,9 @@
 import styles from './book-card.module.scss'
-import type { BookData } from '../../types/types'
 
 import clsx from 'clsx'
 import { BookGallery } from '../images-gallery/images-gallery'
 import { BookInfo } from '../book-info/book-info'
+import type { BookData } from '@shared/lib/types'
 
 import { onest } from '@shared/assets/fonts'
